Review

+++ Assets/Scripts/Messages/CoinsUp.cs
@@ -6,8 +6,8 @@ using CodeControl;
 
 public class CoinsUp : Message
 {
-    public int amount;
-    public CoinsUp(int amount)
+    public double amount;
+    public CoinsUp(double amount)
     {
         this.amount = amount;
     }
